# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstarNews API",
        "description": "Multi-tenant news management API with JWT authentication",
        "contact": {
            "email": "admin@instar.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "description": "Retrieve a paginated list of news. Super admins can see all news, while regular users only see news from their tenants.",
                "operationId": "a03e1a336bc332d11c385fe176757db9",
                "parameters": [
                    {
                        "name": "tenant_uuid",
                        "in": "query",
                        "description": "Filter by tenant UUID (Super Admin only)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "author_uuid",
                        "in": "query",
                        "description": "Filter by author/user UUID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title and content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Create news",
                "description": "Create a new news article. Regular users: uses first tenant automatically. Super Admin: must specify tenant_uuid.",
                "operationId": "a3204d8708e05a4132ae88fec0ac3fe8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Breaking News"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "This is the news content..."
                                    },
                                    "tenant_uuid": {
                                        "description": "Tenant UUID (required for Super Admin, optional for regular users)",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "News created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/news/{uuid}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get a single news",
                "description": "Retrieve a specific news article by UUID",
                "operationId": "f535f9d6d2e99657f918ab5968fc0aa9",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "News UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "News not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Update news",
                "description": "Update an existing news article",
                "operationId": "3c35b60d841573dcaa1f8a62c928b0dd",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "News UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "News updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "News not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete news",
                "description": "Delete a news article (soft delete)",
                "operationId": "e0664727ded8495b3414d477ac38a10b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "News UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "News not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}